Fixing a bug in first_pos_test_results_update due to abs_tmax being an integer. Switched it back to a double and fixed the bug.

+++ src/lib/standard_summary_stats.c
@@ -12,15 +12,15 @@ void std_stats_init(sim_vars* sv, const uint32_t nbinsperunit, bool ngeninfs)
 
   if(sv->pars.timetype==ro_time_first_pos_test_results) {
     stats->nbinsperunit=2*nbinsperunit;
-    stats->tnpersa=stats->npers=(uint32_t)(stats->nbinsperunit*sv->pars.tmax);
+    stats->tnpersa=stats->npers=stats->nbinsperunit*sv->pars.tmax;
     stats->abs_maxnpers=INT32_MAX;
-    stats->abs_tmax=stats->abs_maxnpers/stats->nbinsperunit;
+    stats->abs_tmax=((double)stats->abs_maxnpers)/stats->nbinsperunit;
     stats->first_pos_test_results_time=INFINITY;
     stats->abs_npers=0;
 
   } else {
     stats->nbinsperunit=nbinsperunit;
-    stats->tnpersa=stats->npers=stats->abs_maxnpers=stats->abs_npers=(uint32_t)(stats->nbinsperunit*sv->pars.tmax);
+    stats->tnpersa=stats->npers=stats->abs_maxnpers=stats->abs_npers=stats->nbinsperunit*sv->pars.tmax;
     stats->abs_tmax=sv->pars.tmax;
   }
